requirements.txt
streamlit
groq
chromadb
pypdf
sentence-transformers
python-dotenv